requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
coloredlogs==15.0.1
diskcache==5.6.3
filelock==3.16.1
flatbuffers==24.3.25
fsspec==2024.10.0
huggingface-hub==0.26.5
humanfriendly==10.0
idna==3.10
Jinja2==3.1.4
llama_cpp_python==0.3.5
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.0
onnxruntime==1.20.1
packaging==24.2
protobuf==5.29.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
sentencepiece==0.2.0
setuptools==75.6.0
sympy==1.13.1
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.47.0
typing_extensions==4.12.2
urllib3==2.2.3
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
